refactor:#225 todo 아이템 공통 컴포넌트로 분리
